requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
anthropic>=0.28.0
openai>=1.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.7.0
numpy>=1.26.0
pandas>=2.2.0
pytest>=8.2.0
httpx>=0.27.0
